+++ app-backend/migrations/src/main/scala/migrations/Summary.scala
@@ -159,4 +159,5 @@ object MigrationSummary {
   M158
   M160
   M161
+  M162
 }